+++ KMyMoney.Net.TelegramBot/Commands/Login/LoginCommand.cs
@@ -1,12 +1,11 @@
 using Dropbox.Api;
 using KMyMoney.Net.TelegramBot.Persistence;
 using KMyMoney.Net.TelegramBot.Settings;
-using KMyMoney.Net.TelegramBot.StatusHandlers.Login;
 using Microsoft.Extensions.Options;
 using Telegram.Bot;
 using Telegram.Bot.Types;
 
-namespace KMyMoney.Net.TelegramBot.Commands;
+namespace KMyMoney.Net.TelegramBot.Commands.Login;
 
 public class LoginCommand(
     TelegramBotClientWrapper botWrapper,
